requirements.txt
openai-whisper
moviepy
torch
numpy
soundfile
librosa
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv